Updated pwr.p.test with a higher upper limit for n
Changed the upper limit of n from 5 to 9 digits, for the function to be able to work in situations where n is greater than or equal to 100,000 (lines 52-53). I verified that the program didn't take any longer than it did when the upper limit was 5 digits. For internet-scale testing, it's frequently the case that the effect size is very small, and the number of observations required is larger than 100,000.  For example, for p0=5% and p1=5.15%, the number of observations required is 132367.  Small effects like this over large populations are especially common in e-commerce, search, and advertising.  The legacy function returns an error in such situations because the search space for n is limited to 100,000.  This change increases the search interval for sample size to 10 million, to account for these sorts of scenarios.